refactor(teams): Use parallel lists instead of display_to_short mapping
Instead of building a display_to_short dictionary that we only used for:
1. Getting display names as keys (display_to_short.keys())
2. Getting short names for headers (display_to_short.values())

We now maintain two parallel lists:
- short_names: for the team header
- display_names: for the per_member buffer keys

This is simpler and more direct - we don't need a mapping between them
since we only use display names to match incoming Agno events and
short names for the header. The parallel lists are built together
from the same loop over agent_ids.